[hotfix][runtime] Move PendingWorkerCounter to a separate file.
